Refactor(helm): split install/upgrade into install.go and   upgrade.go with option appliers
Move freshInstall, newInstallAction, and
  isRetryableInstallError into install.go. Extract the inline upgrade
  block from installOrUpgradeChart into a new performUpgrade function
  in upgrade.go.

Introduce applyCommonInstallOptions and
  applyCommonUpgradeOptions free functions that own option-to-action
  wiring per Helm action type. The asymmetry between the two
  (MaxHistory/CleanupOnFail/Recreate live only in the upgrade applier)
   is now visible in code and reflects the Helm SDK shape.

No behavior change. All 240 unit specs remain green.
